api/requirements.txt
fastapi
uvicorn[standard]
python-socketio[asgi]
pydantic
python-multipart
PyJWT
pywebpush
cryptography
